Fix typo: use i64 instead of i32 in requests

+++ src/core/requests/send_message.rs
@@ -17,7 +17,7 @@ pub struct SendMessage {
     #[builder(default)]
     disable_notification: Option<bool>,
     #[builder(default)]
-    reply_to_message_id: Option<i32>,
+    reply_to_message_id: Option<i64>,
     #[builder(default)]
     reply_markup: Option<()>, // TODO: ReplyMarkup enum
 }
